Before I get too carried away, adding progress so far.
The page currently displays our 16x16 grid of evenly sized squares.
The CSS has been adjusted so they are indeed squares, and when text is placed inside it is centred in both dimensions and looks decent and is sized reasonably.
There is also an accent on the edges of each 4x4 tile for easier visibility for players, and a darkening effect on whichever square is being hovered on by the mouse.